Add design documents for deep module refactoring, nesting conventions, raw primitives handling, and sealed ADTs
