+++ .agents/skills/kotlin-standards/history/progression.json
@@ -1,0 +1,6 @@
+{
+  "skill_name": "kotlin-standards",
+  "updated_at": "2026-05-05T01:31:37Z",
+  "benchmarks": [],
+  "case_history": {}
+}
