+++ .agents/skills/kotlin-standards/references/idioms.md
@@ -3,9 +3,14 @@
 Use as a compact expression-style reminder.
 
 Prefer `map`, `mapNotNull`, `flatMap`, `fold`, `associate`, `groupBy`,
-`partition`, `filter`, `takeIf`, `let`, and sealed `when` when they state the
-data flow directly.
+`partition`, `zip`, `buildList`, `filter`, `takeIf`, `let`, and exhaustive
+sealed `when` when they state the data flow directly.
+
+Use higher-order functions when they make the policy clear: how values are
+classified, mapped, folded, ordered, filtered, retried, or wrapped. Name the
+behavior with a local function or well-named parameter when the lambda carries
+business meaning.
 
 Avoid transient `var`s, mutable accumulators, `apply` for non-configuration
-work, `also` chains that hide important side effects, and Java-shaped null
-handling when Kotlin has a clearer construct.
+work, `also` chains that hide important side effects, callback soup, and
+Java-shaped null handling when Kotlin has a clearer construct.
