+++ .github/extensions/kotlin-gradle-loop/extension.mjs
@@ -1,7 +1,8 @@
-import { execFile } from "node:child_process";
-import { dirname, join, resolve } from "node:path";
-import { fileURLToPath } from "node:url";
-import { joinSession } from "@github/copilot-sdk/extension";
+import {execFile} from "node:child_process";
+import {dirname, join, resolve} from "node:path";
+import {fileURLToPath} from "node:url";
+import {joinSession} from "@github/copilot-sdk/extension";
+import {markShadowedExtensionLoaded} from "../_shared/shadowed-skill-state.mjs";
 
 const HERE = dirname(fileURLToPath(import.meta.url));
 const REPO_ROOT = resolve(HERE, "..", "..", "..");
@@ -78,4 +79,5 @@ const tools = [
 ];
 
 const session = await joinSession({ tools });
+markShadowedExtensionLoaded(REPO_ROOT, "kotlin-gradle-loop");
 await session.log("gradle-suite extension ready", { ephemeral: true });
