+++ .github/hooks/hook-state.sh
@@ -28,3 +28,17 @@ PY
     )"
     printf '%s/copilot-hook-required-skills-%s.txt\n' "${TMPDIR:-/tmp}" "${session_key}"
 }
+
+hook_shadowed_extension_state_file() {
+    local repo_root="$1"
+    local session_key
+    session_key="$(
+        python3 - "${repo_root}" <<'PY'
+import hashlib
+import sys
+
+print(hashlib.sha256(sys.argv[1].encode("utf-8")).hexdigest())
+PY
+    )"
+    printf '%s/copilot-hook-shadowed-extensions-%s.txt\n' "${TMPDIR:-/tmp}" "${session_key}"
+}
